refactor: integrate lifecycle into server
